new commit

+++ src/app/@theme/components/header/header.component.ts
@@ -1,10 +1,10 @@
-import { Component, OnDestroy, OnInit } from '@angular/core';
-import { NbMediaBreakpointsService, NbMenuService, NbSidebarService, NbThemeService } from '@nebular/theme';
+import {Component, OnDestroy, OnInit} from '@angular/core';
+import {NbMediaBreakpointsService, NbMenuService, NbSidebarService, NbThemeService} from '@nebular/theme';
 import {NbAuthJWTToken, NbAuthService, NbTokenService} from "@nebular/auth";
-import { UserData } from '../../../@core/data/users';
-import { LayoutService } from '../../../@core/utils';
-import { map, takeUntil } from 'rxjs/operators';
-import { Subject } from 'rxjs';
+import {UserData} from '../../../@core/data/users';
+import {LayoutService} from '../../../@core/utils';
+import {map, takeUntil} from 'rxjs/operators';
+import {Subject} from 'rxjs';
 import {Router} from '@angular/router';
 
 @Component({
@@ -17,6 +17,10 @@ export class HeaderComponent implements OnInit, OnDestroy {
   private destroy$: Subject<void> = new Subject<void>();
   userPictureOnly: boolean = false;
   user: any;
+  admin = {
+    name: 'Admin',
+    picture: 'assets/images/admin.png',
+  };
 
   themes = [
     {
@@ -39,7 +43,8 @@ export class HeaderComponent implements OnInit, OnDestroy {
 
   currentTheme = 'default';
 
-  userMenu =  { title: 'Αποσύνδεση' } ;
+  // userMenu =  { title: 'Αποσύνδεση' } ;
+  userMenu = [{title: 'Log out'}];
 
   constructor(private sidebarService: NbSidebarService,
               // private menuService: NbMenuService,
@@ -50,17 +55,26 @@ export class HeaderComponent implements OnInit, OnDestroy {
               private authService: NbAuthService,
               private tokenService: NbTokenService,
               private router: Router,
-              ) {
+              private menuService: NbMenuService,
+  ) {
+    this.authService.onTokenChange()
+      .subscribe((token: NbAuthJWTToken) => {
+        if (token.isValid()) {
+          this.user = token.getPayload();
+        } else {
+          this.user = this.admin;
+        }
+      });
   }
 
   ngOnInit() {
     this.currentTheme = this.themeService.currentTheme;
 
     // this.userService.getUsers()
-      // .pipe(takeUntil(this.destroy$))
-      // .subscribe((users: any) => this.user = users.nick);
+    //   .pipe(takeUntil(this.destroy$))
+    //   .subscribe((users: any) => this.user = users.nick);
 
-    const { xl } = this.breakpointService.getBreakpointsMap();
+    const {xl} = this.breakpointService.getBreakpointsMap();
     this.themeService.onMediaQueryChange()
       .pipe(
         map(([, currentBreakpoint]) => currentBreakpoint.width < xl),
@@ -70,24 +84,18 @@ export class HeaderComponent implements OnInit, OnDestroy {
 
     this.themeService.onThemeChange()
       .pipe(
-        map(({ name }) => name),
+        map(({name}) => name),
         takeUntil(this.destroy$),
       )
       .subscribe(themeName => this.currentTheme = themeName);
 
-    // this.menuService.onItemClick().subscribe((event) => {
-    //   if (event.item.title === 'Αποσύνδεση') {
-    //     this.tokenService.clear();
-    //     localStorage.clear();
-    //     this.router.navigateByUrl('auth2/login');
-    //   }
-    // });
-  }
-
-  logout(): void {
-    this.tokenService.clear();
-    localStorage.clear();
-    this.router.navigateByUrl('auth2/login');
+    this.menuService.onItemClick().subscribe((event) => {
+      if (event.item.title === 'Log out') {
+        this.tokenService.clear();
+        localStorage.clear();
+        this.router.navigateByUrl('auth2/login');
+      }
+    });
   }
 
   ngOnDestroy() {
@@ -102,7 +110,6 @@ export class HeaderComponent implements OnInit, OnDestroy {
   toggleSidebar(): boolean {
     this.sidebarService.toggle(true, 'menu-sidebar');
     this.layoutService.changeLayoutSize();
-
     return false;
   }
